fix(index.js): rename instances of DcpURL to URL and comment out isURL
Based on feedback from Eddie, DcpURL is a constructor name from
scheduler-v4 changes. The develop branch currently isn't on those
changes so we need to temporarily change the references. FOr the
instances of isURL, I added TODO comments so as to not forget where the
checks are. Some tests still fail because the bundle contains a bug with
the serialization module.