fix(index.js): change logic for wrapping code when evaling in sandbox
Before, any code with a return statement at any indentation would cause
code th be wrapped in an IIFE. But if the code was already wrapped in an
IIFE, the extra wrapping would cause the evaluated code to return
nothing unexpectedly.

So I decided to make the RegEx detect top
level-return statements where wrapping must be needed. Sure, it's mroe
fragile and based on intdentation like in Python, but I'd say the logic
for wrapping is more accurate.