fix(dcp-client.js): patch up dcpConfig Addresses during web bundle load
Wes's patch of correctly representing identities in dcpConfig now
requires that we ensure the identities in the web config are patched up
after the bundle is loaded, not when the config is loaded. Otherwise,
the config would need access to the Address class somehow, before the
bundle has been loaded.