Add minimum support for using `TargetEvent` and its friends
For manually dispatched events, handlers are run synchronously at the
point of dispatch. They do not disrupt our existing strategy for
controlling the event loop.

It's unclear why you also need to the enable the globally available
`dispatchEvent` for using it on the `EventTarget` prototype chain. The
standard doesn't seem to say, but I can't read the DOM standardese.