fix(sandbox): fix same console message detection
The logic for detecting the same console message being sent during
consecutive calls never worked previously due to an array reference
comparison. It would never be truthy due to one of the values always
being a new reference from the rest parameter syntax. That's why there's
logic on the client side as well to detect the duplicate messages.

Even with the duplicate logic fixed in the sandbox, we still always send
the console message data over the network when it isn't necessary.
Deleting the `.message` property suffices for addressing this.

The function/logic was refactored to make it clearer when console
messages are dispatched and to make the relevant functions more focused.

With these changes, the client side logic for propogating the events is
simplified, and will no longer try to also check for duplicate messages.

Related-to: DCP-3998
Refs: https://dstrbtv.slack.com/archives/C7QQPC3J9/p1703013878969179