feat/bug/refactor(bravo-env):get py workFn right before calling slice handler
Before, I erroneously started execution of the Python WorkFunction
before earlier than when the sandbox was ready to process the Py
Slice Handler. This would cause unuseful output when calls to progress
were made inside the Python WorkFucntion (but not in the slice handler)

For ex:
```python
dcp.progress();
dcp.set_slice_handler(lambda x : progress())
```

This didn't work well before, since I was loading the file too early.
Now, I only execute the Python script right before executing the
slice handler. This is when the ability to call `progress` is functional.

---

This is mainly a refactor where code is put into one function instead
of spread out.

---

This also adds the feature of checking if a slice handler is present
in the file which is part of the spec. It will now throw an error
if no slide handler is set in the Python workfunction.