fix(config): fix bypassing config file security check
df5af0a has an issue where setting
`DCP_CLIENT_ALLOW_INSECURE_CONFIGURATION` always caused `dcp-client` to
fail to initialize. I didn't realize that there was code that depended
on the side effect of `checkConfigFileSafePerms` checking the existence
of the file.

Moving the skip of the security checks to be right before the security
related checks, rather than before everything, address that issue.

A test file was added to ensure that issue doesn't get re-introduced
again.