bugifx(sandbox/pyodide*): fix pyodide package loading
NOTE: THIS INCREASES THE MINOR VERSION OF THE WORKTIME

🙃
This bugfix is essentially a new feature since it never worked before.

The bug coveres three main problems
1. Function spelling mistake which caused the package to never be
   loaded properly.
2. Even with (1) fixed, we don't get the correct full package name
   when fetching from the package manager. For instance, if a user
   `import PIL`, it doesn't work since the package is actually called
   `pillow`... So, this change also includes a newly built
   `pyodide-core.js` file which correctly resolves the package name
   based on the import name used.
3. We didn't cast the type of the datum passed to the python slice
   handler to a python object. This seemed to work fine before for
   basic python code, but libraries didn't interact with it well.

As a result of these three bugs fixed, the new pyodide worktime can do
much more than the previous one and therefore should contain a minor
increase since it basically adds new api.

If a tree falls in a forrest, and nobody is around...

If api was broken, and nobody knew it existed, would fixing it be a
a semver patch or semver minor update?
- I'd argue it'd be a minor update!

So this updates the current pyodide worktime to 0.24.0 from 0.23.2.
This is notable because it means we are explicitly deviating from
Pyodiode's semver verisoning. I think this is the better approach
since the Pyodide worktime is its own entity with extra API which
needs to be represented by a versioning system. When there are patch
or API updates to Pyodide itself, we will reflect that into our semver
versions separately I guess...