fix: enable source-map-support for release builds
This'll allow for readable stack traces for anyone that encounters an
error inside the release build of the bundle. There are slight offsets
compared to the source maps for the debug build, but it's an improvement
over dumping the entire bundle when an error is thrown.

It's opt in currently to avoid the library from potentially surprising
users with the patching of `Error.prepareStackTrace` and `process.emit`.

`source-map-support` is declared as a dev dependency given that
`index.js` requires it directly, and that it's already a transitive
dependency of a 'production' dependency, `webpack`.

Refs: https://gitlab.com/Distributed-Compute-Protocol/dcp/-/merge_requests/2573#note_1892717786
Refs: dcp-client!203
Refs: DCP-4151