Emit 'end' event when evaluator connection is TCP RST
When a socket is reset by peer via the `RST` TCP packet, the socket's
`end` event _will not_ be emitted because that strictly means an active
close has started (peer has sent a `FIN`).

But the worker can't proceed anyway, better to let the supervisor know
as early as possible.