bugfix(pyodide): untarring colliding dirs corrupts filesystem
Previously this line used `tar.extractall`; however, this line caused
issues when there happened to be directories with the same path as the
destination the tar is untarred.

For example:
- tar file with `/home/pyodide/text.txt` will corrupt `/home/` and
  `/pyodide/` directories since they already exists in Pyodide by
  default.
- tar file with `/home/will.txt` will corrupt `/home/` since it
  already exists.

This is an issue broadly in Pyodide; I'd have expected it to raise an
exception instead of silently making parts of the filesystem unusable.

This is still an issue in Pyodide as of (0.26.1)...