native event loop: use unix times for nextTimer
nextTimer(n) will cause the ontimer callback to be run when the unix time reaches n (or later). The timers for native were using performance.now for all the timings, which is relative to the start of the program. This caused ontimer to be run _very frequently_ for any large timeout set.